% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add lines to a plot}
\usage{
add_lines(
  x,
  y1,
  y2 = NULL,
  dat = NULL,
  pretty_axis_args = list(pretty = list(n = 5)),
  n = 100,
  f = grDevices::colorRampPalette(c("red", "blue")),
  breaks = NULL,
  cols = NULL,
  add = TRUE,
  output = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector of x values representing the x coordinates of the line.}

\item{y1}{A numeric vector of y values representing the y coordinates of the line.}

\item{y2}{(optional) A numeric vector of values used to colour the line.}

\item{dat}{(optional) A dataframe with columns `x', `y1' and (optionally) `y2' can be provided instead of \code{x}, \code{y1} and \code{y2}.}

\item{pretty_axis_args}{A named list of arguments passed to \code{\link[prettyGraphics]{pretty_axis}}, used to define a sequence of pretty values for colour bar legends (see Examples in \code{\link[prettyGraphics]{add_colour_bar}}).}

\item{n, f}{Colour scheme options (I). \code{n} is the number of bins in \code{y2} to which colours are assigned. Larger numbers produce smoother colour schemes but these can be slower to run. \code{f} is function from which \code{n - 1} colours are extracted.}

\item{breaks, cols}{Colour scheme options (II). \code{breaks} is a sequence of breaks for the colour scheme and \code{cols} is a corresponding vector of colours. (Note that this means there should be one more break than colour.) If supplied, \code{breaks} and \code{cols} silently take priority over \code{n} and \code{f}.}

\item{add}{A logical input that defines whether or not to add the line to the plot.}

\item{output}{(depreciated) A numeric input specifying the output format: \code{1}, list; \code{2}, plot; \code{3}, plot and list (see Value).}

\item{...}{Additional arguments passed to \code{\link[graphics]{lines}}, if only \code{x} and \code{y1} are provided, or \code{\link[graphics]{arrows}} if \code{y2} is also provided.}
}
\value{
The function adds a line to a plot if \code{add = TRUE} and invisibly returns a list with colour bar parameters. The list contains two elements. (1) \code{data_legend} is a dataframe with two columns: `x', a sequence of n values between axis limits (either defined by the user or automatically by \code{\link[prettyGraphics]{pretty_axis}}); and `col', the corresponding colours. This can be passed to \code{\link[prettyGraphics]{add_colour_bar}} to plot a legend. (2) \code{axis_legend} is a list of arguments, generated by \code{\link[prettyGraphics]{pretty_axis}} that can be passed to \code{\link[prettyGraphics]{add_colour_bar}} to add an axis to the legend.
}
\description{
This function is used to add lines to a plot. If only two variables (\code{x} and \code{y1}) are provided, this function collapses to \code{\link[graphics]{lines}}. However, if a third variable is provided (\code{y2}), the line can be coloured by the values of that variable. Colour is incorporated by means of a user-supplied function (\code{f}), from which \code{n - 1} colours are drawn, or a user-supplied sequence of \code{breaks} and corresponding colours (\code{cols}) (e.g., from \code{\link[prettyGraphics]{pretty_cols_brewer}}).
}
\details{
For a line coloured by a third variable, the function works by defining a sequence of 'pretty' values across the range of \code{y2} using \code{\link[prettyGraphics]{pretty_axis}} (unless supplied via \code{breaks}). This series of values can become the labels of a colour bar legend in \code{\link[prettyGraphics]{add_colour_bar}}. Colours are then assigned to each \code{break}, either from a user-supplied function (\code{f}) or a vector of supplied colours (\code{cols}). This approach ensures links between the line drawn and the colour legend added, ensuring both are always 'pretty' (see also \code{\link[prettyGraphics]{add_colour_bar}}).
}
\examples{
#### Generate some example data
x <- 1:1000
y1 <- runif(length(x), 0, 100)
y2 <- x

#### Example (1)
plot(x, y1)
add_lines(x = x, y1 = y1, y2 = y2, n = 100)

}
\seealso{
\code{\link[prettyGraphics]{pretty_axis}}, \code{\link[prettyGraphics]{add_colour_bar}}
}
\author{
Edward Lavender
}
