% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error_envelope.R
\name{add_model_predictions}
\alias{add_model_predictions}
\title{(depreciated) Add model predictions (including fitted values and confidence intervals) to a plot}
\usage{
add_model_predictions(
  x,
  CI = list(),
  fCI = "poly",
  CI_gp = list(col = scales::alpha("lightgrey", 0.8), border = FALSE),
  add_fitted = TRUE,
  fitted_gp = list(col = "black", lwd = 1, lty = 1)
)
}
\arguments{
\item{x}{A numeric vector defining the x values at which fitted values/CIs will be plotted.}

\item{CI}{A named list with fitted values (optional) and lower and upper CIs (i.e. \code{CI$fit}, \code{CI$lowerCI}, \code{CI$upperCI}). This can be created by \code{\link[prettyGraphics]{list_CIs}}.}

\item{fCI}{A character defining the method by which CIs will be added to a plot: as lines (\code{"lines"}) or as a shaded polygon (\code{"poly"}). For large datasets, \code{fCI = "lines"} is much faster.}

\item{CI_gp}{A named list of graphical parameters for CIs. If \code{fCI = "lines"}, then two nested lists can be included to specify the lower and upper CI lines differently, if desired (see Examples).}

\item{add_fitted}{A logical input defining whether or not add the fitted line.}

\item{fitted_gp}{A named list of graphical parameters for the fitted line (passed to \code{\link[graphics]{lines}}).}
}
\value{
The function adds model predictions to a plot.
}
\description{
This function has been renamed. Please use \code{\link[prettyGraphics]{add_error_envelope}} instead.
}
\details{
The function is designed for continuous explanatory variables (i.e., adding regression lines and CIs to a plot). See \code{\link[prettyGraphics]{add_error_bars}} for discrete explanatory variables.
}
\author{
Edward Lavender
}
