% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_moons.R
\name{add_moons}
\alias{add_moons}
\title{Add subplots of the lunar phase to a plot}
\usage{
add_moons(
  side = 3,
  position = 5,
  outer = TRUE,
  nv = 100,
  radius1 = 0.1,
  units = "radians"
)
}
\arguments{
\item{side}{A numeric input that defines the side on to which you would like to add the moons. \code{side = 1}, \code{side = 2}, \code{side = 3}, \code{side = 4} add moons to the bottom, left, top and right axes respectively. The default is to add moons to the top of the plot. This is appropriate if the plot in question is a smooth of lunar phase. If the plot is a lunar phase time-series, \code{side = 2} is more appropriate (see Examples).}

\item{position}{A numeric input that defines the position at which moons will be added to the plot. If \code{side = 1} or \code{side = 3}, this refers to the height at which moons will be added. If \code{side = 2} or \code{side = 4}, this refers to the x value at which moons will be added (see Examples). This should be a single number; i.e. moons are all plotted at the same height (in the case of a smooth plot) or distance along the x axis (in the case of a time-series plot).}

\item{outer}{A logical input that defines whether or not the moons will be drawn beyond the range of the x or y axis. If this is the case, \code{outer = TRUE}; otherwise, set \code{outer = FALSE}.}

\item{nv}{A numeric input that defines the number of vertices that are used to define moon shapes.}

\item{radius1}{A numeric input that defines the radii of the moons added in user units. The default is 0.1.}

\item{units}{A character which defines the units of lunar phase on the existing plot. The default is \code{"radians"}. \code{"degrees"} is the other option.}
}
\value{
Small subplots of the lunar phase (new moon, first quarter, full moon, third quarter, full moon) are added to an existing plot.
}
\description{
This function adds images of the lunar phase to lunar phase plots (for example, a lunar phase time-series or a model-inferred smooth of a response ~ s(lunar phase)).
}
\details{
This function requires the 'lunar' package. Please install this before running this function, using: \code{install.packages("lunar").}
}
\examples{

#### Example (1): Add plots of lunar phase to simulated smooth function:
x <- seq(0, 2*pi, length.out = 100)
y <- (x - 10^2) + 10
plot(x, y)
add_moons(side = 3,
          position = 120,
          outer = FALSE,
          nv = 100,
          radius1 = 0.1,
          units = "radians")

#### Example (2): Add plots to a lunar phase time-series
x <- seq.POSIXt(as.POSIXct("2016-01-01"), as.POSIXct("2016-05-01"), "days")
y <- lunar::lunar.phase(x)
plot(x, y, type = "l")
add_moons(side = 2,
          position = min(x)-1.25e5,
          outer = TRUE,
          nv = 100,
          radius1 = 1e5,
          units = "radians")

}
