% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_map_tools.R
\name{add_north_arrow}
\alias{add_north_arrow}
\title{Add a North arrow to a map}
\usage{
add_north_arrow(
  x0 = "topleft",
  y0 = NULL,
  width,
  mid,
  height,
  type = c("filled head", "filled triangle", "arrow"),
  add_label = NULL,
  col = "black",
  ...
)
}
\arguments{
\item{x0, y0}{A character keyword or doubles that define the arrow's position (see \code{\link[grDevices]{xy.coords}}).}

\item{width, mid, height}{Arrow shape controls. Each argument accepts a double that defines the arrow \code{width}, \code{mid}-point or \code{height}. \code{mid} is only used for \code{type = "filled head"}.}

\item{type}{A character string that defines the arrow type. Currently supported options are (a) \code{"filled head"}, which adds a (filled) arrow head (via \code{\link[graphics]{polygon}}); (b) \code{"filled triangle"}, which adds a (filled) triangle (via \code{\link[graphics]{polygon}}); and (c) \code{"arrow"}, which adds an arrow (via \code{\link[graphics]{arrows}}).}

\item{add_label}{A named list, passed to  \code{\link[graphics]{text}}, to add a label to the plot. \code{add_label = NULL} suppresses this option, \code{add_label = list()} implements default options and a named list customises these.}

\item{col, ...}{Additional graphical parameters passed to \code{\link[graphics]{polygon}} or \code{\link[graphics]{arrows}}.}
}
\value{
The function adds an arrow to a map and invisibly returns a named list with the \code{x} and \code{y} coordinates of the arrow (derived from \code{x0}, \code{y0}, \code{width}, \code{mid} and \code{height} as applicable).
}
\description{
This function adds a North arrow to a map.
}
\details{
Future functionality should include the capacity to control the tilt of the arrow (which is currently vertical) and to add north and east arrows that act as scale bars.
}
\examples{
#### Example (1): Implement function using the default options
raster::plot(dat_gebco)
add_north_arrow()

#### Example (2): Adjust arrow location
raster::plot(dat_gebco)
add_north_arrow(x0 = "topright")
raster::plot(dat_gebco)
add_north_arrow(x0 = -5.55, y = 56.44)

#### Example (3): Adjust arrow type
raster::plot(dat_gebco)
add_north_arrow("topleft")
add_north_arrow("topright", type = "filled triangle")
add_north_arrow("center", type = "filled triangle")
add_north_arrow("right", type = "arrow")

#### Example (4): Adjust arrow type parameters
raster::plot(dat_gebco)
add_north_arrow("topleft", type = "filled triangle", border = "red")
add_north_arrow("center", type = "filled triangle", col = "blue")
add_north_arrow("top", type = "arrow", length = 0.05, lwd = 2)

#### Example (5): Add arrow label
raster::plot(dat_gebco)
add_north_arrow(add_label = list())

}
\author{
Edward Lavender
}
