% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_depreciated}
\alias{check_depreciated}
\title{Check for depreciated arguments passed via \code{...}}
\usage{
check_depreciated(depreciated, ...)
}
\arguments{
\item{depreciated}{A character vector that defines the names of depreciated arguments to a parent function.}

\item{...}{Additional arguments passed to a parent function.}
}
\description{
This function checks for depreciated arguments passed via \code{...} to a parent function and returns a warning if any depreciated arguments have been used.
}
\author{
Edward Lavender
}
\keyword{internal}
