% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{choose_foo_axis}
\alias{choose_foo_axis}
\title{Choose an axis function}
\usage{
choose_foo_axis(at)
}
\arguments{
\item{at}{A sequence of positions at which the axis is drawn. Numbers or times (i.e., dates or POSIXct objects) are supported.}
}
\value{
The function returns either \code{\link[graphics]{axis}}, \code{\link[graphics]{axis.Date}} or \code{\link[graphics]{axis.POSIXct}} depending on the input.
}
\description{
This function returns a function to add an axis to a plot, depending on the input.
}
\author{
Edward Lavender
}
\keyword{internal}
