% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clip_within_range}
\alias{clip_within_range}
\title{Clip a vector to lie within a range}
\usage{
clip_within_range(x, range)
}
\arguments{
\item{x}{A vector, to be clipped.}

\item{range}{A vector of two numbers, used to clip \code{x}.}
}
\value{
A vector, as inputted, solely comprising values within the range specified by \code{range}, inclusive.
}
\description{
This function clips a vector to lie within a range.
}
\author{
Edward Lavender
}
\keyword{internal}
