% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_flapper.R
\docType{data}
\name{dat_flapper}
\alias{dat_flapper}
\title{A sample of flapper skate depth and temperature timeseries}
\format{
A dataframe with 20,161 observations and 4 variables:
\describe{
  \item{timestamp}{A POSIXct object defining the time of each observation (depth/temperature).}
  \item{temp}{A numeric value defining the water temperature (°C) around a given individual at a given time.}
  \item{depth}{A numeric value defining the depth (m) of a given individual at a given time.}
  \item{id}{A factor with two levels ("A" and "B") distinguishing sample timeseries from two different individuals.}
}
}
\source{
Depth timeseries were collected by Marine Scotland Science and Scottish Natural Heritage.
}
\usage{
dat_flapper
}
\description{
A dataset containing a sample of flapper skate (\emph{Dipturus intermedius}) depth and temperature timeseries. Flapper skate are large, Critically Endangered, benthopelagic elasmobranchs. To understand the depth use of individuals within the Loch Sunart to the Sound of Jura Nature Conservation Marine Protected Area, Marine Scotland Science and Scottish Natural Heritage tagged multiple individuals with archival tags, programmed to record depth and temperature at high temporal resolution. \code{dat_flapper} is a small sample of these data including depth (m) and temperature (°C) observations every 2 minutes over a two-week period for two individuals.
}
\keyword{datasets}
