% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{define_lim_init}
\alias{define_lim_init}
\title{Define initial limits from a sequence of numbers.}
\usage{
define_lim_init(x, at = NULL, lim = NULL)
}
\arguments{
\item{x}{A vector which contains user-supplied data.}

\item{at}{(optional) A vector which contains user-supplied axis positions.}

\item{lim}{(optional) A vector of length two which contains user-supplied limits.}
}
\value{
This function returns a vector of two numbers representing initial limits.
}
\description{
. This function defines initial limits for a sequence of numbers.
}
\details{
If the user has supplied limits, this function checks limits to ensure that they are sensible, and flags them with a 'user' attribute (so that user-defined limits are recognisable and remain fixed). If the user has not supplied limits, limits are extracted from axis tick mark positions, if specified, or otherwise from the data provided.
}
\author{
Edward Lavender
}
\keyword{internal}
