% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_time_blocks.R
\name{define_time_blocks}
\alias{define_time_blocks}
\title{Define time blocks (i.e. diel periods or seasons)}
\usage{
define_time_blocks(
  t1 = as.POSIXct("2016-01-01", tz = "UTC"),
  t2 = as.POSIXct("2017-01-01", tz = "UTC"),
  type = "diel",
  type_args = list(),
  to_plot = TRUE,
  col = NULL
)
}
\arguments{
\item{t1}{The start time, in POSIXct format with the time zone specified.}

\item{t2}{The end time, in POSIXct format with the time zone specified.}

\item{type}{A character input specifying the type of block to define. \code{"diel"} or \code{"season"} are currently supported. \code{"diel"} defines a dataframe with \code{date}, \code{time} and \code{level} (1, time of sunrise; 2, time of sunset). \code{"season"} defines a dataframe with \code{date}, \code{time} and \code{level} representing seasons.}

\item{type_args}{A named list of arguments needed to implement \code{type}. This is required for \code{type = "diel"}. A list containing "lon" and "lat" is needed; these arguments are passed to \code{\link[maptools]{sunriset}}.}

\item{to_plot}{A logical input defining whether or not the returned dataframe is to use as an input for plotting (see \code{\link[prettyGraphics]{add_shading_bar}}. If so, the function conducts some processing so that the times are defined exactly along the limits provided by \code{t1} and \code{t2}. This also adjusts the dataframe returned (see Value).}

\item{col}{A vector of colours, one for each factor that are added to the dataframe if \code{to_plot = TRUE}. If \code{col} is \code{NULL}, colours are chosen by default.}
}
\value{
The function returns a dataframe. If \code{to_plot = FALSE}, the dataframe contains 3 columns: \code{date}, \code{time} and \code{level}. If \code{type == "diel"}, levels correspond to day/night; if\code{ type == "season"}, levels correspond to the four seasons. If \code{to_plot = TRUE}, \code{x1}, \code{x2} and \code{col} are returned.
}
\description{
This function defines 'blocks' (i.e., diel periods or seasons) for each day in a time window.
}
\examples{

#### Example (1) define diel blocks between two dates
define_time_blocks(t1 = as.POSIXct("2016-01-01", tz = "UTC"),
                   t2 = as.POSIXct("2016-01-10", tz = "UTC"),
                   type = "diel",
                   type_args = list(lon = 56, lat = -5),
                   to_plot = FALSE)

#### Example (2) define diel blocks between two dates to plot
define_time_blocks(t1 = as.POSIXct("2016-01-01", tz = "UTC"),
                   t2 = as.POSIXct("2016-01-10", tz = "UTC"),
                   type = "diel",
                   type_args = list(lon = 56, lat = -5),
                   to_plot = TRUE,
                   col = c("white", "dimgrey")
                   )

#### Example (3) define seasonal blocks between two dates
define_time_blocks(t1 = as.POSIXct("2016-01-01", tz = "UTC"),
                   t2 = as.POSIXct("2016-01-10", tz = "UTC"),
                   type = "diel",
                   type_args = list(lon = 56, lat = -5),
                   to_plot = FALSE)



}
\author{
Edward Lavender
}
