% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{diff_x}
\alias{diff_x}
\title{Calculate a difference between two values}
\usage{
diff_x(x2, x1)
}
\arguments{
\item{x2}{A number or time.}

\item{x1}{A number or time.}
}
\value{
The function returns the (absolute) difference between \code{x1} and \code{x2}.
}
\description{
Calculate a difference between two values
}
\author{
Edward Lavender
}
\keyword{internal}
