% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_map.R
\name{.crop}
\alias{.crop}
\title{A \code{\link[raster]{crop}} wrapper}
\usage{
.crop(x, ext, layer = c("raster", "polygon", "line", "paths", "points"))
}
\arguments{
\item{x}{A Raster or Spatial* object.}

\item{ext}{An \code{\link[raster]{extent}} object, used to crop \code{x}.}

\item{layer}{A character that defines the spatial layer type. This is used to produce an informative warning message if necessary.}
}
\value{
The function returns \code{x} cropped by \code{ext} or \code{NULL} with a warning if the cropping routine throws an error.
}
\description{
This function is a simple \code{\link[base]{tryCatch}} wrapper for \code{\link[raster]{crop}}. The function attempts to crop a spatial layer, \code{x}, by an \code{\link[raster]{extent}} object. If this proceedure fails, the function throws a warning and returns \code{NULL}; otherwise \code{x} is cropped and returned.
}
\details{
This function is designed to support \code{\link[prettyGraphics]{pretty_map}} and associated funtions.
}
\examples{
\dontrun{
 # Success
.crop(dat_gebco, raster::extent(dat_gebco), layer = "raster")
 # Failure
 .crop(sp::SpatialPoints(cbind(1:10, 1:10)), raster::extent(dat_gebco), layer = "raster")
}
}
\author{
Edward Lavender
}
\keyword{internal}
