% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{empty_list_to_list_null}
\alias{empty_list_to_list_null}
\title{Replace \code{list()} with \code{list(NULL)}}
\usage{
empty_list_to_list_null(list_name, l)
}
\arguments{
\item{list_name}{The name of the list.}

\item{l}{A list.}
}
\value{
A list, as inputted, if the inputted list is not empty; or list(NULL) if the input is an empty list.
}
\description{
This function replaces an empty list with \code{list(NULL)}.
}
\author{
Edward Lavender
}
\keyword{internal}
