% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{extreme_factor}
\alias{extreme_factor}
\alias{min_factor}
\alias{max_factor}
\alias{range_factor}
\title{Factor extremes}
\usage{
min_factor(x)

max_factor(x)

range_factor(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
A factor.
}
\description{
These functions return the minimum/maximum (first/last) levels of a factor.
}
\examples{
x <- factor(LETTERS)
min_factor(x)
max_factor(x)
range_factor(x)
}
\author{
Edward Lavender
}
