% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_notation.R
\name{fractions}
\alias{fractions}
\title{Local implementation of the \code{MASS::fractions()} function}
\source{
This function and the documentation is taken directly from the MASS package (see https://rdrr.io/cran/MASS/src/R/fractions.R). It is defined locally in \code{\link[prettyGraphics]{prettyGraphics}} solely to reduce reliance on non-default packages.
}
\usage{
fractions(x, cycles = 10, max.denominator = 2000, ...)
}
\arguments{
\item{x}{Any object of mode numeric.}

\item{cycles}{The maximum number of steps to be used in the continued fraction approximation process.}

\item{max.denominator}{An early termination criterion.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Find rational approximations to the components of a real numeric object using a standard continued fraction method.
}
\keyword{internal}
