% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implements.R
\name{implement_mtext_args}
\alias{implement_mtext_args}
\title{Implement \code{mtext_args}}
\usage{
implement_mtext_args(mtext_args)
}
\arguments{
\item{mtext_args}{A named list.}
}
\value{
The function adds axes labels to a plot via \code{\link[graphics]{mtext}}.
}
\description{
This function implements the \code{mtext_args} argument in functions.
}
\author{
Edward Lavender
}
