% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implements.R
\name{implement_pretty_axis_args}
\alias{implement_pretty_axis_args}
\title{Implement \code{pretty_axis_args}}
\usage{
implement_pretty_axis_args(pretty_axis_args)
}
\arguments{
\item{pretty_axis_args}{A named list of parameters that are passed to \code{\link[plot.pretty]{pretty_axis}}.}
}
\value{
A list.
}
\description{
This function implements the pretty_axis_args argument in functions. It is not intended to be called directly by the user.
}
\author{
Edward Lavender
}
