% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{is_number}
\alias{is_number}
\title{Check if an object is a number}
\usage{
is_number(x, first = FALSE)
}
\arguments{
\item{x}{An object.}

\item{first}{A logical variable that defines whether or not to select \code{x[[1]]}. This avoids issues with some inputs (e.g., \code{scale(x)} is a matrix).}
}
\value{
A logical value which defines whether or not \code{x} is a number.
}
\description{
This function checks if an input is a number (i.e., an integer or numeric object).
}
\author{
Edward Lavender
}
\keyword{internal}
