% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{is_time}
\alias{is_time}
\title{Check if an object is a time}
\usage{
is_time(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value which defines whether or not \code{x} is a time.
}
\description{
This function checks if an input is a time (i.e., a POSIXct or Date object).
}
\author{
Edward Lavender
}
\keyword{internal}
