% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{list_adjust}
\alias{list_adjust}
\title{Adjust lists to ensure correct mapply loops}
\usage{
list_adjust(l, f = list_depth, side)
}
\arguments{
\item{l}{A list.}

\item{f}{A function used to adjust the list.}

\item{side}{Input to \code{side} argument of \code{\link[prettyGraphics]{pretty_axis}}.}
}
\value{
The function returns an (adjusted) list.
}
\description{
This function adjusts lists to ensure mapply loops over inputted arguments correctly in \code{\link[prettyGraphics]{pretty_axis}}.
}
\author{
Edward Lavender
}
\keyword{internal}
