% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_merge}
\alias{list_merge}
\title{Merge lists accounting for empty lists}
\usage{
list_merge(...)
}
\arguments{
\item{...}{named lists}
}
\description{
This function is a wrapper for \code{\link[rlist]{list.merge}}. The difference is that this function first screens out any empty lists, which cause errors for \code{\link[rlist]{list.merge}}. If there is only one non-empty list, this is returned. Otherwise, \code{\link[rlist]{list.merge}} is used to merge lists in an iterative process. For large lists, this approach will be slower than calling \code{\link[rlist]{list.merge}} directly if there are no empty lists. Both \code{\link[rlist]{list.merge}} and \code{list_merge()} require named lists.
}
\keyword{internal}
