% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pretty.R
\docType{package}
\name{plot.pretty}
\alias{plot.pretty}
\title{plot.pretty: Prettier plots and more detailed data exploration made easier with base R}
\description{
\code{plot.pretty} is an R package designed to make the production of prettier plots and data exploration easier with base R, especially for continuous datasets.
}
\author{
Edward Lavender
}
