% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_cols.R
\name{pretty_cols_brewer}
\alias{pretty_cols_brewer}
\title{Create pretty colour schemes}
\usage{
pretty_cols_brewer(
  zlim,
  buffer = FALSE,
  scheme = "YlOrRd",
  select = 1:8,
  pal = NULL,
  n_breaks = 100,
  rev = FALSE
)
}
\arguments{
\item{zlim}{A numeric range that the colour scheme should span.}

\item{buffer}{A logical input that defines whether or not to 'buffer' breaks along \code{zlim} at either end by the difference between sequential breaks.}

\item{scheme}{A character that defines a colour scheme supported by \code{\link[RColorBrewer]{brewer.pal}}.}

\item{select}{An integer vector that defines which colours to draw from the \code{scheme}.}

\item{pal}{A colour palette function. If supplied \code{scheme} and \code{select} are silently ignored.}

\item{n_breaks}{The number of breaks in the colour scheme (but see \code{buffer}).}

\item{rev}{A logical value that defines whether or not to reverse sampled colours.}
}
\value{
The function returns a named list that contains the z limits (`zlim'), a numeric vector of breaks (`breaks'), the colour palette function (`pal') and the associated vector of colours (`col'). This can be passed to \code{\link[fields]{image.plot}}.
}
\description{
This function is a simple wrapper for \code{\link[RColorBrewer]{brewer.pal}} and \code{\link[grDevices]{colorRampPalette}} or other colour palette functions that generates pretty colour schemes.
}
\details{
This function is designed to work in conjunction with \code{\link[fields]{image.plot}}.
}
\seealso{
\code{\link[prettyGraphics]{pretty_cols_split_heat}}
}
\author{
Edward Lavender
}
