% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_density.R
\name{pretty_density}
\alias{pretty_density}
\title{Pretty probability density plots}
\usage{
pretty_density(
  x,
  f = stats::dgamma,
  param = list(shape = 10, scale = 4),
  pretty_axis_args = list(side = 1:2, pretty = list(n = 5), axis = list(las = TRUE)),
  mtext_args = list(list(side = 1, text = "x", line = 2), list(side = 2, text =
    "Density", line = 3))
)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{f}{A density function.}

\item{param}{A named list of arguments required to evaluate the density function (see Examples).}

\item{pretty_axis_args}{A named list of arguments that are passed to \code{\link[plot.pretty]{pretty_axis}} to define axes.}

\item{mtext_args}{A named list of arguments that are passed to \code{\link[graphics]{mtext}} to define axis labels.}
}
\value{
The function returns a pretty density plot.
}
\description{
The function is designed to produce pretty probability density plots. A rug is added for observed values.
}
\examples{
set.seed(1)
x <- runif(100, 1, 1000)
pretty_density(x = x, f = stats::dgamma, param = list(shape = 10, scale = 4))
pretty_density(x = x, f = stats::dgamma, param = list(shape = 11, scale = 3))
pretty_density(x = x, f = stats::dnorm, param = list(mean = 10, sd = 30))
pretty_density(x = x, f = stats::dnorm, param = list(mean = 10, sd = 30, log = TRUE))

}
\author{
Edward Lavender
}
