% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_hist.R
\name{pretty_hist}
\alias{pretty_hist}
\title{Pretty histograms}
\usage{
pretty_hist(
  x,
  select = "counts",
  pretty_axis_args = list(side = 1:2, pretty = list(n = 10), axis = list(las = TRUE)),
  mtext_args = list(list(side = 1, text = "x", line = 2.5, cex = 1.3), list(side = 2,
    text = "Frequency", line = 2.5, cex = 1.3)),
  ...
)
}
\arguments{
\item{x}{A numeric vector for which to create a histogram.}

\item{select}{A character input (\code{"counts"} or \code{"density"}) specifying the type of histogram to draw.}

\item{pretty_axis_args}{A list of arguments passed to \code{\link[plot.pretty]{pretty_axis}} to define axes.}

\item{mtext_args}{A named list of arguments passed to \code{\link[graphics]{mtext}} to add labels. A nested list is used to control each axis separately.}

\item{...}{other parameters passed to \code{\link[graphics]{hist}} (e.g. if \code{select = "density"}, specify \code{freq = FALSE}).}
}
\value{
The function returns a pretty histogram.
}
\description{
This function uses \code{link[plot.pretty]{pretty_axis}} to make prettier histograms.
}
\examples{

#### Example (1) The default options
set.seed(1)
x <- rnorm(1000, 0, 1)
pretty_hist(x)

#### Example (2) A density plot
pretty_hist(x, select = "density", freq = FALSE)

#### Example (3) Axes can be adjusted via pretty_axis() with pretty_axis_args
pretty_hist(x, pretty_axis_args = list(pretty = list(n = 10), lim = list(range(x), NULL)))

#### Example (4) Axis labels can be adjusted via mtext() and mtext_args()
pretty_hist(x, mtext_args = list(list(side = 1, "x var", line = 2), list(side = 2, "F", line = 2)))

}
\author{
Edward Lavender
}
