% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{pretty_seq}
\alias{pretty_seq}
\title{Define a pretty sequence given data, limits and pretty parameters}
\usage{
pretty_seq(x, lim = NULL, pretty_args = list(n = 5))
}
\arguments{
\item{x}{An object representing some data to be plotted (e.g., x values).}

\item{lim}{(optional) A vector of length two which define the limits within which the sequence should lie.}

\item{pretty_args}{A named list of arguments passed to pretty_x().}
}
\value{
A list comprising 'at', a pretty sequence; and 'lim', limits, which may have been adjusted.
}
\description{
This function defines a pretty sequence given data, limits and parameters.
}
\examples{
pretty_seq(0:10)
pretty_seq(0:10, lim = c(0, 9))
pretty_seq(seq.Date(as.Date("2016-01-01"), as.Date("2016-01-12"), 1))
pretty_seq(seq.Date(as.Date("2016-01-01"), as.Date("2016-01-12"), 1), pretty_args = list(n = 3))

}
\author{
Edward Lavender
}
