% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_ts_mat.R
\name{pretty_ts_mat}
\alias{pretty_ts_mat}
\title{Visualise the within and between day variation in a timeseries}
\usage{
pretty_ts_mat(
  x,
  y,
  res,
  t1_units = function(x) x/60,
  t2_units = function(x) as.character(x),
  retain_orientation = TRUE,
  xtick_every_n = 5,
  ytick_every_n = 60/res,
  xlab = "Time (days)",
  ylab = "Time (hours)",
  return_mat = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A vector of timestamps in POSIXct (\code{\link[base]{DateTimeClasses}}) format.}

\item{y}{A numeric vector of values of the response variable.}

\item{res}{A number which defines the resolution (in minutes) between sequential observations.}

\item{t1_units}{A function which defines the units of the axis of within-day variation. Since \code{res} is in minutes, \code{function(x) x} results in units of minutes, while the default \code{function(x) x/60} results in units of hours.}

\item{t2_units}{A function which defines the format of the axis of between-day variation. The function takes in objects of class \code{\link[base]{Date}} and returns axis labels. The default option is to retain dates in YYYY-MM-DD format, but other formats (e.g., Julian day) can be expressed via this option.}

\item{retain_orientation}{A logical input which defines whether to plot time of day against time of year (\code{TRUE}) or vice-versa (\code{FALSE}).}

\item{xtick_every_n}{A numeric input which defines the spacing between sequential tick marks for the x axis (see \code{\link[prettyGraphics]{pretty_mat}}).}

\item{ytick_every_n}{A numeric input which defines the spacing between sequential tick marks for the x axis (see \code{\link[prettyGraphics]{pretty_mat}}).}

\item{xlab}{The x axis label.}

\item{ylab}{The y axis label.}

\item{return_mat}{A logical input which defines whether or not to return the matrix plotted.}

\item{verbose}{A logical input which defines whether or not to print messages to the console. This can be useful in monitoring function progress, especially for large datasets.}

\item{...}{Additional arguments passed to \code{\link[prettyGraphics]{pretty_mat}} for plot customisation.}
}
\value{
The function returns a plot and, if requested, a matrix of time of day (in user-specified units but hours by default) x time of year (in user-specified units but days by default).
}
\description{
This function produces a 2-dimensional time plot of a response variable. The surface of the plot shows the response variable in relation to time of day on one axis (usually the x axis) and time of year (specifically, the date) on another axis (usually the y axis). The function was motivated by the need to visualise how the depth of aquatic animals changes over the course of the day and how these patterns change over the course of the year (e.g., as in Teo et al., 2013).
}
\examples{
#### Define an example dataframe
dfa <- dat_flapper[dat_flapper$id == "A", ]

#### Example (1): Visualise plot for different response variables
# ... sampled at 2 minutes resolution
pretty_ts_mat(dfa$timestamp, dfa$depth, res = 2)
pretty_ts_mat(dfa$timestamp, dfa$temp, res = 2)

#### Example (2): Adjust the units on the time of day axis via t1_units
pretty_ts_mat(dfa$timestamp, dfa$depth,
                  res = 2,
                  t1_units = function(x) x,
                  ylab = "Time (mins)")

#### Example (3): Adjust the format of the time of year axis via t2_units
# Julian day
pretty_ts_mat(dfa$timestamp, dfa$depth,
                  res = 2,
                  t2_units = function(x) lubridate::yday(x))
# day-month
pretty_ts_mat(dfa$timestamp, dfa$depth,
              res = 2,
              t2_units = function(x) format(x, "\%d-\%m"))

#### Example (4): Adjust the number of tick marks via xtick_every_n and ytick_every_n
pretty_ts_mat(dfa$timestamp, dfa$depth,
                  res = 2,
                  t1_units = function(x) x,
                  ytick_every = 120,
                  ylab = "Time (mins)")
# Other customisation is implemented via ... (see ?prettyGraphics::pretty_mat)

}
\references{
Teo, S.L., Sandstrom, P.T., Chapman, E.D., Null, R.E., Brown, K., Klimley, A.P., Block, B.A., 2013. Archival and acoustic tags reveal the post-spawning migrations, diving behav- ior, and thermal habitat of hatchery-origin Sacramento River steelhead kelts (\emph{Oncorhynchus mykiss}). Environ. Biol. Fish 96, 175–187.
}
\seealso{
The plot is produced by \code{\link[prettyGraphics]{pretty_mat}}.
}
\author{
Edward Lavender
}
