% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{pretty_x}
\alias{pretty_x}
\title{Create a regular pretty sequence}
\usage{
pretty_x(obj, x, ...)
}
\arguments{
\item{x}{A vector of numbers or times.}

\item{...}{Other arguments passed to \code{\link[base]{pretty}} or \code{\link[lubridate]{pretty_dates}}.}

\item{An}{object.}
}
\value{
The function returns a series of pretty numbers or times.
}
\description{
This function creates a regular pretty sequence from numbers or times.
}
\seealso{
\code{link[prettyGraphics]{pretty_seq}} is an extension of this function.
}
\author{
Edward Lavender
}
\keyword{internal}
