% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyGraphics-package.R
\docType{package}
\name{prettyGraphics}
\alias{prettyGraphics}
\title{prettyGraphics: Prettier plots and more detailed data exploration made easier with base R}
\description{
\code{prettyGraphics} is an R package designed to make the production of prettier plots and data exploration easier with base R, especially for continuous datasets.
}
\author{
Edward Lavender
}
