% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_xy.R
\name{pull_xy}
\alias{pull_xy}
\title{Pull x and y coordinates from supported R objects for pretty plotting functions}
\usage{
pull_xy(x, y = NULL)
}
\arguments{
\item{x}{x coordinates or an object from which x and y coordinates can be extracted.}

\item{y}{y coordinates.}
}
\value{
A list with 'x' and 'y' coordinates, extracted from the object \code{x} or as inputted, as well as 'xlab' and 'ylab' labels, if manual labels are desirable.
}
\description{
This function pulls 'x' and 'y' coordinates from some R objects. This is implemented within \code{\link[prettyGraphics]{pretty_axis}} and \code{\link[prettyGraphics]{pretty_plot}} to create axes for multiple object types. Suggested x and y axis labels are also returned: these are NULL if automatically defined labels are deemed appropriate or character strings which can be used to replace automatically defined labels.
}
\author{
Edward Lavender
}
\keyword{internal}
