% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{seq_extend}
\alias{seq_extend}
\title{Extend a regular sequence within limits}
\usage{
seq_extend(x, lim)
}
\arguments{
\item{x}{A sequence... Need to handle dates etc....}

\item{lim}{A vector of two which specify the lower and upper limits.}
}
\value{
The function returns a sequence.
}
\description{
This function extends a regular sequence in both directions towards user-defined limits.
}
\examples{
# extend_seq(2:4, c(0, 10))
# extend_seq(2:4, c(0, 4))
# x <- seq.Date(as.Date("2016-01-01"), as.Date("2016-01-10"), 1)
# extend_seq(x, as.Date(c("2015-12-10", "2016-01-20")))
}
\author{
Edward Lavender
}
\keyword{internal}
