% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{seq_x}
\alias{seq_x}
\title{Create a regular sequence between two values}
\usage{
seq_x(x1, x2, units)
}
\arguments{
\item{x1}{The first number/time in the sequence.}

\item{x2}{The last number/time in the sequence.}

\item{units}{The distance between sequential values in the sequence.}
}
\value{
The function returns a regular sequence between two values.
}
\description{
This function creates a sequence between two values (numbers or times).
}
\author{
Edward Lavender
}
\keyword{internal}
