% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_map_tools.R
\name{summarise_by_lat}
\alias{summarise_by_lat}
\title{Summarise a \code{\link[raster]{raster}} by latitude}
\usage{
summarise_by_lat(x, stat = mean)
}
\arguments{
\item{x}{A \code{\link[raster]{raster}}.}

\item{stat}{A function that summarises the values of \code{x} across latitudinal bands.}
}
\value{
The function returns a dataframe with the latitudinal coordinates of each band (`y') and the summary (`stat'),
}
\description{
This function summarises the values of a \code{\link[raster]{raster}} across latitudinal bands using a user-supplied function, such as \code{\link[base]{mean}}.
}
\details{
This routine is implemented within \code{\link[prettyGraphics]{add_profile_lat}}.
}
\examples{
d <- summarise_by_lat(dat_gebco)
utils::head(d)
}
\author{
Edward Lavender
}
