% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_tables.R
\name{tidy_write}
\alias{tidy_write}
\title{Write `tidy' tables to file}
\usage{
tidy_write(x, file, quote = FALSE, sep = ",", row.names = FALSE, na = "-", ...)
}
\arguments{
\item{x, file, quote, sep, row.names, na, ...}{Arguments passed to \code{\link[utils]{write.table}}.}
}
\value{
The function writes a table to file.
}
\description{
This is a wrapper function for \code{\link[utils]{write.table}} with `tidier' defaults. This is designed to write tables to file that can be easily copied into word processors, such as Microsoft Word.
}
