% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{units_x}
\alias{units_x}
\title{Define units within \code{\link[prettyGraphics]{pretty_axis}}}
\usage{
units_x(iunits, x)
}
\arguments{
\item{iunits}{A number or character. If \code{NULL}, the function defines suitable units based on the object class (see \code{x} and Value). Otherwise, the function returns \code{iunits} unchanged.}

\item{x}{An object.}
}
\value{
If \code{iunits} is \code{NULL}, the function returns units = 5 for number inputs and units = "auto" for time inputs. Otherwise, the function returns \code{iunits} as inputted.
}
\description{
This function defines suitable default units (i.e., distances between sequential observations in a regular sequence) for a given input type (numbers or times).
}
\author{
Edward Lavender
}
\keyword{internal}
