% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines.R
\name{add_lines}
\alias{add_lines}
\title{Add lines to a plot}
\usage{
add_lines(
  x,
  y1,
  y2 = NULL,
  dat = NULL,
  pretty_axis_args = list(pretty = list(n = 5)),
  n = 100,
  f = grDevices::colorRampPalette(c("red", "blue")),
  output = 3,
  ...
)
}
\arguments{
\item{x}{A numeric input of x values representing the x coordinates of the line to be plotted.}

\item{y1}{A numeric input of y values representing the y coordinates of the lines to be plotted.}

\item{y2}{(optional) A numeric input of values which will be used to colour the lines plotted.}

\item{dat}{(optional) A dataframe with columns 'x', 'y1' and (optionally) 'y2' can be provided instead of \code{x}, \code{y1} and \code{y2}.}

\item{pretty_axis_args}{A list of arguments passed to \code{\link[prettyGraphics]{pretty_axis}}. This is used to define a sequence of pretty values, between the limits of which \code{n} colours are drawn. Modifying \code{pretty_axis_args} predominantly affects the colour bar legend that is added to plots (see Examples in \code{\link[prettyGraphics]{add_colour_bar}}).}

\item{n}{(optional) The number of bins in \code{y2} for which a colour is assigned. Larger numbers produce smoother colour schemes but these can be slower to run.}

\item{f}{(optional) A function used to define the colours that are used to colour the line according to \code{y2}.}

\item{output}{A numeric input specifying the output format: \code{1}, list; \code{2}, plot; \code{3}, plot and list (see Value).}

\item{...}{Additional arguments passed to \code{\link[graphics]{lines}}, if only \code{x} and \code{y1} are provided, or \code{\link[graphics]{arrows}} if \code{y2} is also provided.}
}
\value{
A line and/or a list. The list contains two elements. (1) \code{data_legend} is a dataframe with two columns: 'x', a sequence of n values across limits (either defined by the user or automatically by \code{\link[prettyGraphics]{pretty_axis}}) values; and 'col', referring to corresponding colours. This can be passed to \code{\link[prettyGraphics]{add_colour_bar}} to plot a legend. (2) \code{axis_legend} is a list of arguments, generated by \code{\link[prettyGraphics]{pretty_axis}} that can be passed to \code{\link[prettyGraphics]{add_colour_bar}} to add an axis to the legend.
}
\description{
This function is used to add lines to a plot. If only two variables (\code{x} and \code{y1}) are provided, this function collapses to \code{\link[graphics]{lines}}. However, if a third variable is provided (\code{y2}), the line can be coloured by the values of that variable. This can aid inferences regarding associations between two variables and a third variable (or indeed, help emphasise relationships between a response and a single explanatory variable with colour). In this case, the function works by defining a sequence of 'pretty' values across the range of \code{y2} using \code{\link[prettyGraphics]{pretty_axis}}. This range of values can become the labels of a colour bar legend in \code{\link[prettyGraphics]{add_colour_bar}}. Within this sequence of values, the function defines \code{n} regularly spaced values and, using a user-specified function, a colour is defined for each of these values. The line is plotted by matching each value of \code{y2} with the closest value in this sequence. The larger \code{n}, the smoother the colour scheme (at some computational cost). The approach ensures links between the line drawn and the colour legend added, ensuring both are always 'pretty' (see also \code{\link[prettyGraphics]{add_colour_bar}}).
}
\examples{

#### Generate some example data
x <- 1:1000
y1 <- runif(length(x), 0, 100)
y2 <- x

#### Example (1)
plot(x, y1)
add_lines(x = x, y1 = y1, y2 = y2, n = 100)

}
\seealso{
\code{\link[prettyGraphics]{pretty_axis}}, \code{\link[prettyGraphics]{add_colour_bar}}
}
\author{
Edward Lavender
}
