% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implements.R
\name{implement_mtext_args}
\alias{implement_mtext_args}
\title{Implement \code{mtext_args}}
\usage{
implement_mtext_args(mtext_args)
}
\arguments{
\item{mtext_args}{A named list.}
}
\value{
A list.
}
\description{
This function implements the \code{mtext_args} argument in functions. It is not intended to be called directly by the user.
}
\author{
Edward Lavender
}
