% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implements.R
\name{implement_pretty_axis_args}
\alias{implement_pretty_axis_args}
\title{Implement \code{pretty_axis_args}}
\usage{
implement_pretty_axis_args(x, pretty_axis_args)
}
\arguments{
\item{x}{A list of of coordinates (see \code{\link[prettyGraphics]{pretty_axis}}).}

\item{pretty_axis_args}{A named list of parameters that are passed to \code{\link[prettyGraphics]{pretty_axis}}.}
}
\value{
A list, returned by \code{\link[prettyGraphics]{pretty_axis}}.
}
\description{
This function implements the \code{pretty_axis_args} argument in functions.
}
\author{
Edward Lavender
}
