% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{is_number}
\alias{is_number}
\title{Check if an object is a number}
\usage{
is_number(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value which defines whether or not \code{x} is a number.
}
\description{
This function checks if an input is a number (i.e., an integer or numeric object).
}
\author{
Edward Lavender
}
\keyword{internal}
