% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_support.R
\name{pretty_labels}
\alias{pretty_labels}
\title{Define pretty labels}
\usage{
pretty_labels(x, at)
}
\arguments{
\item{x}{An object, such as a numeric vector.}

\item{at}{A numeric vector of axis positions.}
}
\value{
A vector of labels, of the same length as axis tick marks (\code{at}).
}
\description{
This function defines pretty axis labels from an object.
}
\details{
For factors, factor levels at positions specified by \code{at} are taken as labels. For numbers, \code{\link[plot.pretty]{add_lagging_point_zero}} and \code{\link[plot.pretty]{sci_notation}} are implemented as necessary to define pretty labels.
}
\author{
Edward Lavender
}
\keyword{internal}
