% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_smooths.R
\name{pretty_smooth_2d}
\alias{pretty_smooth_2d}
\title{Pretty two-dimensional smooths}
\usage{
pretty_smooth_2d(
  x,
  view = NULL,
  xlim,
  ylim,
  pretty_axis_args = list(side = 1:4, axis = list(list(), list(), list(labels = FALSE),
    list(labels = FALSE))),
  add_xy = NULL,
  add_rug_x = NULL,
  add_rug_y = NULL,
  return_list = FALSE,
  ...
)
}
\arguments{
\item{x}{An mgcv model (see \code{\link[mgcv]{vis.gam}}).}

\item{view}{A character vector of two variables (see \code{\link[mgcv]{vis.gam}}).}

\item{xlim, ylim, pretty_axis_args}{Axis control arguments. \code{xlim} and \code{ylim} control x and y axis limits via \code{\link[prettyGraphics]{pretty_axis}}. If not supplied, \code{xlim} and \code{ylim} are defined as the range across each corresponding variable. To use default 'pretty' limits instead, specify \code{xlim = NULL} and \code{ylim = NULL}. Additional axis customisation is implemented by passing a named list of arguments to \code{\link[prettyGraphics]{pretty_axis}} via \code{pretty_axis_args}.}

\item{add_xy}{A named list of arguments, passed to \code{\link[graphics]{points}}, to add observations to the plot. \code{add_xy = NULL} suppresses this option, \code{add_xy = list()} implements default arguments and a named list customises these.}

\item{add_rug_x, add_rug_y}{Named list of arguments, passed to \code{\link[graphics]{rug}}, to add observed values of the variables defined in \code{view} to the plot. \code{add_rug_* = NULL} suppresses this option, \code{add_rug_*} implements default arguments and a named list customises these.}

\item{return_list}{A logical input which defines whether or not to return the list produced by \code{\link[prettyGraphics]{pretty_axis}}.}

\item{...}{Additional arguments passed to \code{\link[mgcv]{vis.gam}}, excluding \code{"plot.type"} (since only contour plots are supported by this function) and \code{"axes"}.}
}
\value{
The function returns a contour plot of the predictions of a generalised addition model for the two variables defined in \code{view}.
}
\description{
This is function plots pretty two-dimensional smooths as a contour plot.
}
\details{
At present, the function is simply a wrapper for \code{\link[mgcv]{vis.gam}} with the additional flexibility provided by the \code{\link[prettyGraphics]{pretty_axis}} function and the \code{add_xy}, \code{add_rug_x} and \code{add_rug_y} arguments.
}
\examples{
#### Simulate example data and fit model (following ?mgcv::vis.gam examples)
set.seed(0)
n    <- 200
sig2 <- 4
x0   <- runif(n, 0, 1)
x1   <- runif(n, 0, 1)
x2   <- runif(n, 0, 1)
y    <- x0^2 + x1 * x2 + runif(n, -0.3, 0.3)
g    <- mgcv::gam(y ~ s(x0, x1, x2))

#### Example (1): Contour plot using default options
pretty_smooth_2d(g, view = c("x1", "x2"))

#### Example (2): Customise axes via xlim, ylim and pretty_axis_args
# Use xlim and ylim
pretty_smooth_2d(g, view = c("x1", "x2"),
                 xlim = c(0, 1),
                 ylim = c(0, 1))
# Use pretty_axis_args
pretty_smooth_2d(g, view = c("x1", "x2"),
                 pretty_axis_args = list(side = 1:4))

#### Example (3): Add observed data
# Specify list() to use default options
pretty_smooth_2d(g, view = c("x1", "x2"),
                 add_xy = list())
# Customise addition of observed data
pretty_smooth_2d(g, view = c("x1", "x2"),
                 add_xy = list(pch = ".'", cex = 5))

#### Example (4): Add rugs for the x and y variables
# Use default options
pretty_smooth_2d(g, view = c("x1", "x2"),
                 add_rug_x = list(),
                 add_rug_y = list())
# Customise options
pretty_smooth_2d(g, view = c("x1", "x2"),
                 add_rug_x = list(col = "grey"),
                 add_rug_y = list(col = "grey"))

#### Example (5): Pass additional options to mgcv::vis.gam() via ...
pretty_smooth_2d(g, view = c("x1", "x2"),
                 main = "", xlab = "x1[...]", ylab = "x2[...]",
                 color = "gray",
                 contour.col = "red")

}
\author{
Edward Lavender
}
