% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_xy.R
\name{pull_xy}
\alias{pull_xy}
\title{Pull x and y coordinates from supported R objects for \code{\link[plot.pretty]{pretty_axis}}}
\usage{
pull_xy(x, y = NULL)
}
\arguments{
\item{x}{x coordinates or an object from which x and y coordinates can be extracted.}

\item{y}{y coordinates.}
}
\value{
A list with 'x' and 'y' coordinates, extracted from the object \code{x} or as inputted.
}
\description{
This function pulls 'x' and 'y' coordinates from some R objects. This is implemented within \code{\link[plot.pretty]{pretty_plot}}, \code{\link[plot.pretty]{pretty_axis}} to create axes for multiple object types.
}
\author{
Edward Lavender
}
\keyword{internal}
