% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_notation.R
\name{sci_notation}
\alias{sci_notation}
\title{Tidy scientific notation}
\usage{
sci_notation(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector of expression objects that can be added to a plot.
}
\description{
Reformat numbers in scientific notation, translating the default 'e' notation used in base R to the 'x10' format more suitable for publication quality plots. All numbers in a vector are translated into expression objects with the 'x10' notation which can be added to plots. For consistency, any numbers without 'e' in a vector are treated similarly (e.g. \eqn{1} becomes \eqn{1 \times 10^0})
}
\examples{

#### Example (1):
x <- seq(1e-10, 1e10, by = 1e9)
y <- runif(length(x), 0, 100)
xtidy <- sci_notation(x)
plot(x, y, axes = FALSE)
axis(side = 1, at = x, labels = xtidy, pos = 0, las = 2)
axis(side = 2, at = seq(0, 100, by = 10), pos = 1e-10)

}
\author{
Edward Lavender
}
